requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
psycopg2-binary==2.9.10
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==1.1.0
google-cloud-documentai==2.29.0
google-auth==2.29.0
google-api-python-client==2.155.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyJWT==2.8.0
pydantic==2.5.3
google-generativeai==0.8.3
